In SliderProps type definition, replace `void` types with `null`. (#8102)
With TypeScript compiler option ‘strictNullChecks’ enabled, the value `null` cannot be assigned to type `void`. This means, for instance, that it’s impossible to pass a `null` value for prop `step`. I don’t expect that substituting `null` for `void` in these type definitions should cause anyone difficulty; because these are optional props, a consumer of this component should be able to pass `null` or `undefined` safely, whether or not the compiler is using strict null checks.